requirements.txt
fastapi
uvicorn
jinja2
python-multipart
chromadb
pypdf
passlib
bcrypt
python-dotenv
openai
